+++ src/components/Heading/Heading.tsx
@@ -9,7 +9,6 @@ import React, { forwardRef } from "react";
 
 import Link from "../Link/Link";
 import Text from "../Text/Text";
-import useNYPLBreakpoints from "../../hooks/useNYPLBreakpoints";
 
 export const headingSizesArray = [
   "display1",
@@ -130,7 +129,7 @@ export const Heading: ChakraComponent<
         noSpace,
         url,
       });
-      const { isLargerThanMobile } = useNYPLBreakpoints();
+
       // Combine native base styles with any additional styles.
       // This is used in the `Hero` and `Notification` components.
       const asHeading: any = finalLevel;
@@ -175,86 +174,26 @@ export const Heading: ChakraComponent<
         contentToRender
       );
 
-      /** *********************************************************************
-       * The syntax for responsive styles is not working properly for fontSize
-       * in the theme object, so logic for the responsive styles has been
-       * written here.
-       * *********************************************************************/
-
-      /** The default style used to map the `level` value to the corresponding
-       * `size` value. */
-      const defaultRoot = "heading";
-
-      /** The heading size index that acts as a separator between the smaller
-       * and larger sizing styles of the overline and subtitle elements. This
-       * demarcation is purely based on design and aesthetics. */
-      const overlineSubtitleSizeDemarcation = 2;
-
-      /**  Most variant values have a number at the end, so let's remove the
-       * last character from that value and see what's left. */
-      const variantRoot = variant.slice(0, -1);
-
-      /** The `level` values should map to a corresponding `size` value. If the
-       * root is "h", then reassign the root to the `defaultRoot`. This value
-       * will be used to build the style object. */
-      const finalRoot = variantRoot === "h" ? defaultRoot : variantRoot;
-
-      /** The new heading styles use a number with the variant style to indicate
-       * which style should be used. For example, heading1, heading2, and so on.
-       * If that number is set, we'll need it later. Let's grab the last
-       * character in string now, so we can use later in the code. In fact,
-       * let's grab that character and type it as an integer. */
-      const sizeIndex = parseInt(variant.at(-1));
-
-      /** The 2023 typography styles call for the Heading component to be
-       * responsive and set differing font-size values for desktop and mobile.
-       * We can tell if the new styles are being used if the size index is in
-       * fact a number. If it is a number, let's go ahead and setup the
-       * responsive styles. If it is not a number, then the deprecated styles
-       * are being used and we don't need to worry about responsive font-size
-       * styles.
-       * */
-      const responsiveStyles = !isNaN(sizeIndex)
-        ? isLargerThanMobile
-          ? {
-              fontSize: `desktop.heading.${finalRoot}${sizeIndex}`,
-            }
-          : { fontSize: `mobile.heading.${finalRoot}${sizeIndex}` }
-        : undefined;
-
-      /** If the overline element is rendered, we'll also need responsive styles
-       * for that. */
-      const overlineSize = !isNaN(sizeIndex)
-        ? sizeIndex <= overlineSubtitleSizeDemarcation
-          ? "overline1"
-          : "overline2"
-        : undefined;
-      const overlineFontSize = !isNaN(sizeIndex)
-        ? sizeIndex <= overlineSubtitleSizeDemarcation
-          ? isLargerThanMobile
-            ? "desktop.overline.overline1"
-            : "mobile.overline.overline1"
-          : isLargerThanMobile
-          ? "desktop.overline.overline2"
-          : "mobile.overline.overline2"
-        : undefined;
-
-      /** If the subtitle element is rendered, we'll also need responsive styles
-       * for that. */
-      const subtitleSize = !isNaN(sizeIndex)
-        ? sizeIndex <= overlineSubtitleSizeDemarcation
-          ? "subtitle1"
-          : "subtitle2"
-        : undefined;
-      const subtitleFontSize = !isNaN(sizeIndex)
-        ? sizeIndex <= overlineSubtitleSizeDemarcation
-          ? isLargerThanMobile
-            ? "desktop.subtitle.subtitle1"
-            : "mobile.subtitle.subtitle1"
-          : isLargerThanMobile
-          ? "desktop.subtitle.subtitle2"
-          : "mobile.subtitle.subtitle2"
-        : undefined;
+      /** The sizes of the overline and subtitle elements are based on the
+       * current heading variant. The heading variants that reqwuire the larger
+       * text styles are listed in the usesLargerTextSizes array. If the name of
+       * the current heading variant is included in that array, then the size
+       * prop of the overline and subtitle elements will be set to "overline1"
+       * and "subtitle1" respectively.
+       */
+      const usesLargerTextSizes = [
+        "display1",
+        "heading1",
+        "heading2",
+        "h1",
+        "h2",
+      ];
+      const overlineSize = usesLargerTextSizes.includes(variant)
+        ? "overline1"
+        : "overline2";
+      const subtitleSize = usesLargerTextSizes.includes(variant)
+        ? "subtitle1"
+        : "subtitle2";
 
       /** The styles that should be applied to the outer-most wrapper of the
        * Heading component. */
@@ -268,11 +207,9 @@ export const Heading: ChakraComponent<
         overline || subtitle
           ? {
               ...styles.base,
-              ...responsiveStyles,
             }
           : {
               ...styles.base,
-              ...responsiveStyles,
               ...wrapperStyles,
             };
 
@@ -285,7 +222,6 @@ export const Heading: ChakraComponent<
               mb="xxs"
               role="paragraph"
               size={overlineSize}
-              fontSize={overlineFontSize}
             >
               {overline}
             </Text>
@@ -309,7 +245,6 @@ export const Heading: ChakraComponent<
               noSpace
               role="paragraph"
               size={subtitleSize}
-              fontSize={subtitleFontSize}
             >
               {subtitle}
             </Text>