+++ src/components/Text/textChangelogData.ts
@@ -10,15 +10,18 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
-    date: "2023-12-07",
-    version: "2.1.3",
-    type: "Update",
-    affects: ["Styles"],
-    notes: ["Chakra 2.8 update."],
-  },
-  {
     date: "Prerelease",
     version: "Prerelease",
+    type: "Update",
+    affects: ["Styles"],
+    notes: [
+      "Chakra 2.8 update.",
+      "Updates the component's theme file to use native Chakra responsive styles for the responsive font sizes.",
+    ],
+  },
+  {
+    date: "2023-12-07",
+    version: "2.1.3",
     type: "Update",
     affects: ["Accessibility"],
     notes: [