+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Updates
 
 - Updates the `DatePicker`'s `initialDate` and `initialDateTo` props to accept an empty string.
+- Updates the `Heading` component to use native Chakra responsive styles to handle the responsive text size for the component's internal `heading`, `overline` and `subtitle` elements. This also resolves the flashing font size bug that is most noticeable on slower internet connections.
+- Updates the `Text` component to use native Chakra responsive styles to handle the responsive text size for the `"overline1"`, `"overline2"`, `"subtitle1"` and `"subtitle2"` variants.
 
 ## 2.1.3 (December 7, 2023)